+++ lib/route/origin_security.go
@@ -3,6 +3,8 @@ package route
 import (
 	"context"
 	"net/http"
+
+	log "github.com/sirupsen/logrus"
 )
 
 type matchedDomainKey int
@@ -14,9 +16,16 @@ type matchedDomainKey int
 // OriginSecurity will store the matching domain in the http.Request's Context. Use MatchedDomain
 // to retrieve the value in later logic.
 func OriginSecurity(domains []Domain) SecurityHandler {
+	var validDomains []string
+	for _, d := range domains {
+		validDomains = append(validDomains, d.String())
+	}
+	logger := log.WithFields(log.Fields{"validDomains": validDomains})
+
 	return func(h http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			domain := FindDomain(r.Header.Get("Origin"), domains)
+			origin := r.Header.Get("Origin")
+			domain := FindDomain(origin, domains)
 			if domain != nil {
 				ctx := r.Context()
 				ctx = context.WithValue(ctx, matchedDomainKey(0), domain)
@@ -25,6 +34,11 @@ func OriginSecurity(domains []Domain) SecurityHandler {
 				return
 			}
 
+			if len(origin) == 0 {
+				logger.Debug("Request origin is missing")
+			} else {
+				logger.WithFields(log.Fields{"origin": origin}).Debug("Request origin is invalid")
+			}
 			w.WriteHeader(http.StatusForbidden)
 			w.Write([]byte("Origin is not a trusted host."))
 		})